+++ src/cpp/embedded_file.cpp
@@ -28,16 +28,12 @@ namespace poppler
 PYBIND11_MODULE(embedded_file, m)
 {
     py::class_<embedded_file>(m, "embedded_file")
-        .def("checksum", &embedded_file::checksum)
-        .def("checksum_bytes", [](const embedded_file &self){
-            auto &&data = self.checksum();
-            return py::bytes(&data[0], data.size());
+        .def("checksum", [](const embedded_file& self){
+            return py::bytes(&self.checksum()[0], data.size());
         })
         .def("creation_date", &embedded_file::creation_date)
-        .def("data", &embedded_file::data)
-        .def("data_bytes", [](const embedded_file &self){
-            auto &&data = self.data();
-            return py::bytes(&data[0], data.size());
+        .def("data", [](const embedded_file& self){
+            return py::bytes(&self.data()[0], data.size());
         })
         .def("description", &embedded_file::description)
         .def("is_valid", &embedded_file::is_valid)