+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Next release
 
 - Updated pybind11 to 2.6.2
 - Tested with Python 3.9 and Poppler 20.11.0
+- EmbeddedFile data() and checksum() now return bytes (See :pr:`32`) -- by Bence Cs
 
 0.2.2 (2020-10-03)
 ------------------